+++ lib/representable/hash.rb
@@ -17,6 +17,7 @@ module Representable
 
     module ClassMethods
       def from_hash(*args, &block)
+        warn "[Representable] Deprecation warning: Class methods ::from_hash, ::from_json, ::from_xml, etc. will be removed in 2.0. Please use Model.new.from_json instead. Keep smiling."
         create_represented(*args, &block).from_hash(*args)
       end
     end