+++ lib/representable/pipeline.rb
@@ -1,0 +1,13 @@
+module Representable
+  # Allows to implement a pipeline of filters where a value gets passed in and the result gets
+  # passed to the next callable object.
+  #
+  # Note: this is still experimental.
+  class Pipeline < Array
+    include Uber::Callable
+
+    def call(context, value, *args)
+      inject(value) { |memo, block| block.call(memo, *args) }
+    end
+  end
+end