+++ CHANGES.md
@@ -8,9 +8,9 @@ If you need to inherit common methods to all inline decorators, use **** TODO: d
 * `Config` no longer is a Hash, it's API is limited to a few methods like `#collect`, `#<<`, etc.
 * Moved `Representable::ClassMethods::Declarations` to `Representable::Declarative`.
 * Moved `Representable::ClassMethods` to `Representable::Declarative`.
-
 * Fixed: Inline decorators now work with `inherit: true`.
 * Remove `:extend` in combination with inline representer. The `:extend` option is no longer considered. Include the module directly into the inline block.
+* Deprecated class methods `::from_json` and friends. Use the instance version on an instance.
 
 # 1.8.5